+++ services/src/main/java/org/keycloak/protocol/oidc/DefaultTokenExchangeProvider.java
@@ -302,11 +302,11 @@ public class DefaultTokenExchangeProvider implements TokenExchangeProvider {
             }
         }
 
-        if (targetClient.isConsentRequired()) {
-            event.detail(Details.REASON, "audience requires consent");
-            event.error(Errors.CONSENT_DENIED);
-            throw new CorsErrorResponseException(cors, OAuthErrorException.INVALID_CLIENT, "Client requires user consent", Response.Status.BAD_REQUEST);
-        }
+//        if (targetClient.isConsentRequired()) {
+//            event.detail(Details.REASON, "audience requires consent");
+//            event.error(Errors.CONSENT_DENIED);
+//            throw new CorsErrorResponseException(cors, OAuthErrorException.INVALID_CLIENT, "Client requires user consent", Response.Status.BAD_REQUEST);
+//        }
 
         if (!targetClient.equals(client) && !AdminPermissions.management(session, realm).clients().canExchangeTo(client, targetClient)) {
             event.detail(Details.REASON, "client not allowed to exchange to audience");